requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
numpy>=1.26
opencv-contrib-python>=4.8
# Optional: install only if you have a working C++ toolchain
# face_recognition>=1.3
